fix after tests

+++ migrations/accrual/000002_create_reward_rules_table.up.sql
@@ -1,6 +1,6 @@
 -- Таблица правил вознаграждений
 CREATE TABLE IF NOT EXISTS reward_rules (
-    match        TEXT      PRIMARY KEY,
-    reward       BIGINT    NOT NULL,
-    reward_type  TEXT      NOT NULL CHECK (reward_type IN ('%', 'pt'))
+    match        TEXT             PRIMARY KEY,
+    reward       NUMERIC(10,2)    NOT NULL,
+    reward_type  TEXT             NOT NULL CHECK (reward_type IN ('%', 'pt'))
 );
